+++ Models/BooksModels.cs
@@ -6,7 +6,7 @@ public class BooksModels
     public string Title { get; set; }
     public string Author { get; set; }
     public string Year { get; set; }
-    public decimal Money { get; set; }
+    public double Money { get; set; }
     public string Gender { get; set; }
     public string Description { get; set; }
 }
